% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infoMetrics.R
\name{extract_infoMetrics}
\alias{extract_infoMetrics}
\title{Extract Various Metrics from INFO Field of vcf File}
\usage{
extract_infoMetrics(expandedVCFfile)
}
\arguments{
\item{expandedVCFfile}{object of ExpandedVCF class}
}
\value{
dataframe object containing the requested information for the INFO field of the vcf file
}
\description{
Extract metrics from the INFO field of a vcf file. Returns a data.frame containing one row for each variant and columns containing information about the type of variant, read depth, number of reads supporting the variant allele, number of reads supporting the reference allele, and the percentage of reads supporting the variant
}
\examples{
\dontrun{extract_infoMetrics(expandedVCFfile)}

}
