% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rowData.R
\name{extract_rowData}
\alias{extract_rowData}
\title{Extract rowData from ExpandedVCF file}
\usage{
extract_rowData(expandedVCFfile)
}
\arguments{
\item{expandedVCFfile}{object of ExpandedVCF class}
}
\value{
GRanges object containing genomic coordinates for each variant and in the metadata columns Reference (REF) and Alternate (ALT) alleles, as well as quality scores (QUAL) and filter status
}
\description{
Extract the rowData from ExpandedVCF as GRanges object and convert seqlevels/chromsome names to UCSC style
}
\examples{
\dontrun{extract_rowData(expandedVCFfile)}
}
