% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variantConsequence.R
\name{extract_variantConsequence}
\alias{extract_variantConsequence}
\title{Annotate Coding Variants by Biological Consequence}
\usage{
extract_variantConsequence(rowData)
}
\arguments{
\item{rowData}{GRanges object containing genomic coordinates for each variant and Reference (REF) and Alternate (ALT) alleles}
}
\value{
a tibble object containing information about the consequence of each variant in the CONSEQUENCE column along with additional information of interest about each variant including the reference allele codon and amino acid (REFCODON, REFAA), the alternative allele codon and amino acid (VARCODON, VARAA) and the variant name in the ExAC style
}
\description{
Extracts consequences for coding variants, returning only the most deleterious consequence per variant if there are multiple consequences.
}
\examples{
\dontrun{extract_variantConsequence(rowData)}

}
