% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variantLocation.R
\name{extract_variantLocation}
\alias{extract_variantLocation}
\title{Annotate Variants Based on Genomic Feature Location}
\usage{
extract_variantLocation(rowData)
}
\arguments{
\item{rowData}{GRanges object containing genomic coordinates for each variant and Reference (REF) and Alternate (ALT) alleles}
}
\value{
tibble object containing information about the location of each variant with respect to genomic features and if located within a gene additionally the Gene ID number for that annotation
}
\description{
Extracts location of variants within genomic features
}
\examples{
\dontrun{extract_variantLocation(rowData)}

}
