% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importExpandedVCF.R
\name{importExpandedVCF}
\alias{importExpandedVCF}
\title{Import vcf file and generate ExpandedVCF object}
\usage{
importExpandedVCF(vcfFile, genome = "hg19")
}
\arguments{
\item{vcfFile}{character}

\item{genome}{character}
}
\value{
ExpandedVCF
}
\description{
Import vcf file and generate ExpandedVCF object
}
\examples{
importExpandedVCF("myVariants.vcf", "hg19")
}
