% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varAnnoPrototype-package.R
\docType{package}
\name{varAnnoPrototype-package}
\alias{varAnnoPrototype}
\alias{varAnnoPrototype-package}
\title{varAnnoPrototype: A Prototype Variant Annotation Tool}
\description{
vcf files are complex and can be difficult to interpret. 
    varAnnoPrototype is a prototype variant annotation tool that extracts some 
    relevant information from vcf files in addition to annotating the variants 
    contained within the file to the location and type of genomic feature in 
    which they are located as well as their biological consequence. This 
    package contains a set of functions to read in a vcf file, extract the 
    relevant information, perform the genomic annotations, and output a 
    dataframe containing the result which can be written to file, viewed in 
    external software, and shared with others.
}
\author{
\strong{Maintainer}: R.A. Greenstein \email{rgreenstein@wustl.edu}

}
\keyword{internal}
